Changed audioWarp buffer name handling
Buffer name change does does not unlock the markers or affect markers. Buffer name changes are deferred (low) in order to take priority over restored snapshot at load.

In timeWarpInterface.js, replaced bang() with mgraphics.redraw() in drawbuffer in order no to resend the markers when the buffer is changed.